Refactor position.animateTo to position.animate.to
