+++ src/application/internal/ApplicationState.js
@@ -72,7 +72,7 @@ export class ApplicationState
    /**
     * Restores a saved application state returning the data. Several optional parameters are available
     * to control whether the restore action occurs silently (no store / inline styles updates), animates
-    * to the stored data, or simply sets the stored data. Restoring via {@link Position.animateTo} allows
+    * to the stored data, or simply sets the stored data. Restoring via {@link AnimationImpl.to} allows
     * specification of the duration, easing, and interpolate functions along with configuring a Promise to be
     * returned if awaiting the end of the animation.
     *
@@ -213,12 +213,12 @@ export class ApplicationState
                // Return a Promise with saved data that resolves after animation ends.
                if (async)
                {
-                  return application.position.animateTo(data.position, { duration, ease, interpolate }).finished.then(
+                  return application.position.animate.to(data.position, { duration, ease, interpolate }).finished.then(
                    () => application);
                }
                else  // Animate synchronously.
                {
-                  application.position.animateTo(data.position, { duration, ease, interpolate });
+                  application.position.animate.to(data.position, { duration, ease, interpolate });
                }
             }
             else
